Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-pannernode-interface/panner-automation-position-expected.txt

CONSOLE MESSAGE: line 159: TypeError: undefined is not an object (evaluating 'panner.positionX.setValueAtTime')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'panner.positionX.setValueAtTime')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [0: 1-channel inverse rolloff: 1]  
